参数判断

+++ src/main/java/cn/com/njdhy/muscle/biceps/controller/sys/RoleCtl.java
@@ -66,6 +66,9 @@ public class RoleCtl {
     @RequestMapping("/{id}")
     public Result queryById(@PathVariable String id) {
 
+        if (ObjectUtils.isEmpty(id)){
+            return Result.error("500", "参数不能为空");
+        }
         SysRole model = sysRoleService.queryById(id);
 
         if (ObjectUtils.isEmpty(model)) {
@@ -87,7 +90,9 @@ public class RoleCtl {
 
         try {
             // 校验参数 todo
-
+            if (ObjectUtils.isEmpty(sysRole)){
+                return Result.error("500", "角色信息不能为空");
+            }
             // 执行入库操作
             sysRoleService.insertRoleInfo(sysRole);
         } catch (ApplicationException e) {
@@ -113,7 +118,12 @@ public class RoleCtl {
         try {
             // 校验参数
             // TODO: 2018/3/14
-
+            if (ObjectUtils.isEmpty(sysRole)){
+                return Result.error("500", "角色信息不能为空");
+            }
+            if (ObjectUtils.isEmpty(sysRole.getId())){
+                return Result.error("500", "角色id不能为空");
+            }
             // 执行修改
             sysRoleService.updateRoleInfo(sysRole);
         } catch (RuntimeException e) {
@@ -138,6 +148,9 @@ public class RoleCtl {
 
         try {
             // 校验参数 todo
+            if (ObjectUtils.isEmpty(ids)){
+                return Result.error("500", "角色id不能为空");
+            }
             sysRoleService.deleteByIds(ids);
         } catch (ApplicationException e) {
             LOGGER.error(e.getMsg());
